Move getBrandedImage to lib/utilities for general use
