+++ src/app/lib/utilities/getBrandedImage/index.test.js
